adds home page test
moves home tests out of django rendered and adds dfp targeting test
